Add Execute method for operation with options
